working biomes + tekstures
all is kinda working

+++ env_biomes.h
@@ -3,8 +3,8 @@
 #include "Core.h"
 
 class env_biomes {
-	int humidity;
-	int temperature;
+	double humidity;
+	double temperature;
 public:
 	void generate_biomes(sprite_params** data, int width, int height);
 };
